modified:   home/lucas/.config/i3/config 	modified:   home/lucas/.config/i3blocks/scripts/btc 	new file:   home/lucas/.config/polybar/anews.py 	new file:   home/lucas/.config/polybar/btc.sh 	modified:   home/lucas/.config/polybar/config 	modified:   home/lucas/.config/polybar/config.dracula 	new file:   home/lucas/.config/polybar/config.dracula2 	new file:   home/lucas/.config/polybar/config.galdino2 	new file:   home/lucas/.config/polybar/config.szorfein 	new file:   home/lucas/.config/polybar/config.szorfein2 	new file:   home/lucas/.config/polybar/crypto.py 	modified:   home/lucas/.config/polybar/launch.sh 	deleted:    home/lucas/.config/polybar/launch2.sh 	modified:   home/lucas/.config/polybar/weather.py 	modified:   home/lucas/.config/rofi/scripts/configs 	modified:   home/lucas/bin/matebg.sh 	modified:   home/lucas/bin/nodisplay.sh

+++ home/lucas/bin/nodisplay.sh
@@ -4,17 +4,20 @@
 #declare -A apps
 # Array numerica
 declare -a apps
+declare -a element
 
 homedir="${HOME}/.local/share/applications"
 sysdir="/usr/share/applications"
-apps=("cinnamon-" "qt4" "mate-" "mpv")
+apps=("gnome" "cinnamon" "qt4" "mate" "zenmap" "mpv.desktop" "links.desktop" "avahi-discover.desktop" "bssh.desktop" "bvnc.desktop")
 todos=($(find $sysdir $homedir -type f | egrep .desktop))
 nd='NoDisplay'
 
-#function uso {}
+function uso {
+	echo "Uso: $(basename $0) -rsa"
+}
 
 function nodisp {
-	if [ $1 = '-r' ]; then
+	if [ "$1" = '-r' ]; then
         if [ -f $2 ]; then
 			grep 'NoDisplay' $2 1> /dev/null
         	if [ $? = 0 ]; then
@@ -33,10 +36,10 @@ function nodisp {
 	fi
 }
 
-if [ $1 ]; then
-	if [ $1 = '-s' ]; then
-		if [ $2 ]; then
-			arquivo=$(grep -r $2 /usr/share/applications/ | awk -F':' '{print $1}' | uniq | head -n1)
+if [ "$1" ]; then
+	if [ "$1" = "-s" ]; then
+		if [ "$2" ]; then
+			arquivo=$(grep -ri $@ /usr/share/applications/ | awk -F':' '{print $1}' | uniq | head -n1)
 			if [ ! -z $arquivo ]; then
 				echo $arquivo
 				echo $arquivo | xclip -r -selection c
@@ -44,8 +47,7 @@ if [ $1 ]; then
 				echo "Nenhum arquivo encontrado."
 			fi
 		fi
-
-	elif [ $1 = '-r' ]; then
+	elif [ "$1" = '-r' ]; then
 		if [ -f "${homedir}/$(basename $2)" ]; then
 			nodisp -r "${homedir}/$(basename $2)"
 		fi
@@ -55,25 +57,29 @@ if [ $1 ]; then
 		cp "${sysdir}/$(basename $1)" "${homedir}/"
 		nodisp "${homedir}/$(basename $1)"
 	fi
-	exit 0
+
+elif [ "$1" = '-a' ]; then
+	for app in ${apps[@]}; do
+		if [ -f $app ]; then
+			echo "Procurando o regex ${app}..."
+	    	for element in "${todos[@]}"; do
+	        	if [[ $element = *$app* ]]; then
+	            	echo $element
+	            	nome=$(basename $element)
+					echo "O $nome coincide com ${app}..."
+
+	    			if [ ! -f "${homedir}/${nome}" ]; then
+	    		   		echo "$nome não existe em ${homedir}"
+	    		   		if [ -f "${sysdir}/${nome}" ]; then
+	    		       		echo "$nome existe em ${sysdir}, copiando para ${homedir}/${nome}"
+	    		       		cp "${sysdir}/${nome}" "${homedir}/"
+	    		   		fi
+	    			fi
+					[ -f "${homedir}/${nome}" ] && nodisp "${homedir}/${nome}"
+	        	fi
+	    	done
+    	fi
+	done
+else
+	uso
 fi
-
-for app in ${apps[@]}; do
-	echo "Procurando o regex ${app}..."
-    for element in "${todos[@]}"; do
-        if [[ $element = *$app* ]]; then
-            nome=$(basename $element)
-			echo "O $nome coincide com ${app}..."
-
-    		if [ ! -f "${homedir}/${nome}" ]; then
-    		   echo "$nome não existe em ${homedir}"
-    		   if [ -f "${sysdir}/${nome}" ]; then
-    		       echo "$nome existe em ${sysdir}, copiando para ${homedir}/${nome}"
-    		       cp "${sysdir}/${nome}" "${homedir}/"
-    		   fi
-    		fi
-			[ -f "${homedir}/${nome}" ] && nodisp "${homedir}/${nome}"
-
-        fi
-    done
-done
